+++ src/components/common/badges/InterestBadge.tsx
