+++ src/components/kid/home/Pending.stories.tsx
@@ -1,9 +1,9 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import Margin from '../layout/Margin';
+import Margin from '../../layout/Margin';
 import Pending from './Pending';
 
 export default {
-  title: '자녀 홈/Pending',
+  title: 'Kid/home/Pending',
   component: Pending,
   argTypes: { onClick: { action: 'event' } },
   decorators: [
