+++ src/components/kid/home/Summary.stories.tsx
@@ -1,10 +1,10 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import Margin from '../layout/Margin';
+import Margin from '../../layout/Margin';
 import Summary from './Summary';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: '자녀 홈/Summary',
+  title: 'Kid/home/Summary',
   component: Summary,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   decorators: [
