+++ src/components/layout/Layout.tsx
@@ -2,7 +2,6 @@ import { Outlet, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import { media } from '../../lib/styles/theme';
 import { useMediaQuery } from 'react-responsive';
-import { borderRadius } from 'polished';
 
 function Layout() {
   const { pathname } = useLocation();
